full res unit types, refactor enum definitions
