+++ types/enums/DevelopmentTypes.ts
@@ -7,3 +7,20 @@ export const DevelopmentTypes = {
   newBuild: 'New build',
   notKnown: 'Not known',
 };
+
+type DevelopmentTypeKeys = keyof typeof DevelopmentTypes;
+
+type GenericDevelopmentType<TKey extends DevelopmentTypeKeys> = {
+  value: TKey;
+  description: (typeof DevelopmentTypes)[TKey];
+};
+
+type DevelopmentTypeMap = {
+  [K in DevelopmentTypeKeys]: GenericDevelopmentType<K>;
+};
+
+/**
+ * @id #DevelopmentType
+ * @description Development types
+ */
+export type DevelopmentType = DevelopmentTypeMap[keyof DevelopmentTypeMap];
