+++ types/enums/HousingProviders.ts
@@ -9,3 +9,20 @@ export const GLAHousingProviders = {
   publicAuthority: 'Other public authority',
   selfBuild: 'Self-build',
 };
+
+type GLAHousingProviderKeys = keyof typeof GLAHousingProviders;
+
+type GenericGLAHousingProvider<TKey extends GLAHousingProviderKeys> = {
+  value: TKey;
+  description: (typeof GLAHousingProviders)[TKey];
+};
+
+type GLAHousingProviderMap = {
+  [K in GLAHousingProviderKeys]: GenericGLAHousingProvider<K>;
+};
+
+/**
+ * @id #GLAHousingProvider
+ * @description Housing provider categories tracked by the Greater London Authority (GLA)
+ */
+export type GLAHousingProvider = GLAHousingProviderMap[keyof GLAHousingProviderMap];
