+++ types/enums/ResidentialUnitTypes.ts
@@ -1,4 +1,4 @@
-export const NationalResidentialUnitTypes = {
+export const UKResidentialUnitTypes = {
   cluster: 'Cluster flat',
   flat: 'Flat',
   house: 'House',
@@ -23,6 +23,40 @@ export const GLAResidentialUnitTypes = {
 };
 
 export const ResidentialUnitTypes = {
-  ...NationalResidentialUnitTypes,
+  ...UKResidentialUnitTypes,
   ...GLAResidentialUnitTypes,
 };
+
+type UKResidentialUnitTypeKeys = keyof typeof UKResidentialUnitTypes;
+
+type GenericUKResidentialUnitType<TKey extends UKResidentialUnitTypeKeys> = {
+  value: TKey;
+  description: (typeof UKResidentialUnitTypes)[TKey];
+};
+
+type UKResidentialUnitTypeMap = {
+  [K in UKResidentialUnitTypeKeys]: GenericUKResidentialUnitType<K>;
+};
+
+/**
+ * @id #UKResidentialUnitType
+ * @description Residential unit types tracked throughout the UK
+ */
+export type UKResidentialUnitType = UKResidentialUnitTypeMap[keyof UKResidentialUnitTypeMap];
+
+type GLAResidentialUnitTypeKeys = keyof typeof GLAResidentialUnitTypes;
+
+type GenericGLAResidentialUnitType<TKey extends GLAResidentialUnitTypeKeys> = {
+  value: TKey;
+  description: (typeof GLAResidentialUnitTypes)[TKey];
+};
+
+type GLAResidentialUnitTypeMap = {
+  [K in GLAResidentialUnitTypeKeys]: GenericGLAResidentialUnitType<K>;
+};
+
+/**
+ * @id #GLAResidentialUnitType
+ * @description Residential unit types tracked by the Greater London Authority (GLA)
+ */
+export type GLAResidentialUnitType = GLAResidentialUnitTypeMap[keyof GLAResidentialUnitTypeMap];
