+++ types/enums/TenureTypes.ts
@@ -1,4 +1,4 @@
-export const NationalTenureTypes = {
+export const UKTenureTypes = {
   AHO: 'Affordable home ownership',
   MH: 'Market housing',
   other: 'Other',
@@ -25,6 +25,40 @@ export const GLATenureTypes = {
 };
 
 export const TenureTypes = {
-  ...NationalTenureTypes,
+  ...UKTenureTypes,
   ...GLATenureTypes,
 };
+
+type UKTenureTypeKeys = keyof typeof UKTenureTypes;
+
+type GenericUKTenureType<TKey extends UKTenureTypeKeys> = {
+  value: TKey;
+  description: (typeof UKTenureTypes)[TKey];
+};
+
+type UKTenureTypeMap = {
+  [K in UKTenureTypeKeys]: GenericUKTenureType<K>;
+};
+
+/**
+ * @id #UKTenureType
+ * @description Tenure types tracked throughout the UK
+ */
+export type UKTenureType = UKTenureTypeMap[keyof UKTenureTypeMap];
+
+type GLATenureTypeKeys = keyof typeof GLATenureTypes;
+
+type GenericGLATenureType<TKey extends GLATenureTypeKeys> = {
+  value: TKey;
+  description: (typeof GLATenureTypes)[TKey];
+};
+
+type GLATenureTypeMap = {
+  [K in GLATenureTypeKeys]: GenericGLATenureType<K>;
+};
+
+/**
+ * @id #GLATenureType
+ * @description Tenure types tracked by the Greater London Authority (GLA)
+ */
+export type GLATenureType = GLATenureTypeMap[keyof GLATenureTypeMap];
