+++ types/schema/File.ts
@@ -1,4 +1,4 @@
-import {FileTypes} from '../enums/FileTypes';
+import {FileType} from '../enums/FileTypes';
 
 /**
  * @id #File
@@ -9,20 +9,3 @@ export interface File {
   type: FileType[];
   description?: string;
 }
-
-type FileTypeKeys = keyof typeof FileTypes;
-
-type GenericFileType<TKey extends FileTypeKeys> = {
-  value: TKey;
-  description: (typeof FileTypes)[TKey];
-};
-
-type FileTypeMap = {
-  [K in FileTypeKeys]: GenericFileType<K>;
-};
-
-/**
- * @id #FileType
- * @description Types of planning documents and drawings
- */
-export type FileType = FileTypeMap[keyof FileTypeMap];
