+++ types/schema/Metadata.ts
@@ -1,6 +1,6 @@
-import {DateTime, URL, UUID} from './../utils';
-import {FileType} from './File';
-import {QuestionMetaData} from './Responses';
+import { FileType } from '../enums/FileTypes';
+import { DateTime, URL, UUID } from './../utils';
+import { QuestionMetaData } from './Responses';
 
 /**
  * @id #DigitalPlanningMetadata
