+++ types/schema/data/Proposal.ts
@@ -1,11 +1,16 @@
+import { BuildingRegulation } from '../../enums/BuildingRegulations';
+import { DevelopmentType } from '../../enums/DevelopmentTypes';
+import { GLAHousingProvider } from "../../enums/HousingProviders";
 import {
-  GLAOpenSpaceTypes,
-  GLAOpenSpaceDesignations,
+  OpenSpaceDesignation,
+  OpenSpaceType
 } from '../../enums/OpenSpaces';
-import {GLAProtectedSpaceDesignations} from '../../enums/ProtectedSpaces';
-import {ProjectTypes} from '../../enums/ProjectTypes';
-import {Area, Date} from '../../utils';
-import {GeoBoundary, Materials} from './shared';
+import { ProjectType } from '../../enums/ProjectTypes';
+import { ProtectedSpaceDesignation } from '../../enums/ProtectedSpaces';
+import { GLAResidentialUnitType, UKResidentialUnitType } from '../../enums/ResidentialUnitTypes';
+import { GLATenureType } from '../../enums/TenureTypes';
+import { Area, Date } from '../../utils';
+import { GeoBoundary, Materials } from './shared';
 
 /**
  * @id #Proposal
@@ -106,8 +111,8 @@ export interface BaseProposal {
   };
   units?: {
     residential: {
-      development: string; // enum
-      type: string; // enum
+      development: DevelopmentType;
+      type: UKResidentialUnitType;
       bedrooms: number;
       identicalUnits: number;
     }[];
@@ -268,96 +273,23 @@ export interface ProposalDates {
   completion?: Date;
 }
 
-type ProjectTypeKeys = keyof typeof ProjectTypes;
-
-type GenericProjectType<TKey extends ProjectTypeKeys> = {
-  value: TKey;
-  description: (typeof ProjectTypes)[TKey];
-};
-
-type ProjectTypeMap = {
-  [K in ProjectTypeKeys]: GenericProjectType<K>;
-};
-
-/**
- * @id #ProjectType
- * @description Planning project types
- */
-export type ProjectType = ProjectTypeMap[keyof ProjectTypeMap];
-
-type OpenSpaceTypeKeys = keyof typeof GLAOpenSpaceTypes;
-
-type GenericOpenSpaceType<TKey extends OpenSpaceTypeKeys> = {
-  value: TKey;
-  description: (typeof GLAOpenSpaceTypes)[TKey];
-};
-
-type OpenSpaceTypeMap = {
-  [K in OpenSpaceTypeKeys]: GenericOpenSpaceType<K>;
-};
-
-/**
- * @id #OpenSpaceType
- * @description Types of natural open spaces
- */
-export type OpenSpaceType = OpenSpaceTypeMap[keyof OpenSpaceTypeMap];
-
-type OpenSpaceDesignationKeys = keyof typeof GLAOpenSpaceDesignations;
-
-type GenericOpenSpaceDesignation<TKey extends OpenSpaceDesignationKeys> = {
-  value: TKey;
-  description: (typeof GLAOpenSpaceDesignations)[TKey];
-};
-
-type OpenSpaceDesignationMap = {
-  [K in OpenSpaceDesignationKeys]: GenericOpenSpaceDesignation<K>;
-};
-
-/**
- * @id #OpenSpaceDesignation
- * @description Designations of natural open spaces
- */
-export type OpenSpaceDesignation =
-  OpenSpaceDesignationMap[keyof OpenSpaceDesignationMap];
-
-type ProtectedSpaceDesignationKeys = keyof typeof GLAProtectedSpaceDesignations;
-
-type GenericProtectedSpaceDesignation<
-  TKey extends ProtectedSpaceDesignationKeys,
-> = {
-  value: TKey;
-  description: (typeof GLAProtectedSpaceDesignations)[TKey];
-};
-
-type ProtectedSpaceDesignationMap = {
-  [K in ProtectedSpaceDesignationKeys]: GenericProtectedSpaceDesignation<K>;
-};
-
-/**
- * @id #ProtectedSpaceDesignation
- * @description Designations of natural protected spaces
- */
-export type ProtectedSpaceDesignation =
-  ProtectedSpaceDesignationMap[keyof ProtectedSpaceDesignationMap];
-
 interface GLARetainedUnit {
   bedrooms: number;
-  tenure: string; // enum
-  type: string; // enum
+  tenure: GLATenureType;
+  type: GLAResidentialUnitType;
   identicalUnits: number;
 }
 
 interface GLALostUnit extends GLARetainedUnit {
   habitableRooms: number;
-  compliance: string; // enum
-  provider: string; // enum
+  compliance: BuildingRegulation[];
+  provider: GLAHousingProvider;
   area: Area;
   sheltered: boolean;
   olderPersons: boolean;
 }
 
 interface GLAGainedUnit extends GLALostUnit {
-  development: string; // enum
+  development: DevelopmentType;
   garden: boolean;
-  sheltered: boolean;
 }
