+++ types/schema/data/shared.ts
@@ -16,3 +16,18 @@ export type GeoBoundary = {
   site: GeoJSON;
   area: Area;
 };
+
+export type Entity = {
+  name: string;
+  description?: string;
+  source: PlanningDataSource | OSRoadsSource;
+};
+
+type PlanningDataSource = {
+  text: 'Planning Data';
+  url: URL;
+};
+
+type OSRoadsSource = {
+  text: 'Ordnance Survey MasterMap Highways';
+};
